requirements.txt
customtkinter
psutil
pillow
packaging
